+++ python/pyxel/examples/07_snake.py
@@ -88,7 +88,7 @@ class Snake:
         if pyxel.btn(pyxel.KEY_Q):
             pyxel.quit()
 
-        if pyxel.btnp(pyxel.KEY_R) or pyxel.btnp(pyxel.GAMEPAD1_BUTTON_X):
+        if pyxel.btnp(pyxel.KEY_R) or pyxel.btnp(pyxel.GAMEPAD1_BUTTON_START):
             self.reset()
 
     def update_direction(self):